update design: created header comp and making it responsive
