+++ src/App.css
@@ -497,7 +497,7 @@
     text-decoration: none;
   }
 
-  .Mobile-Menu, .Mobile-Menu-Close {
+  .Mobile-Menu {
     display: none;
   }
